requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.0
chromadb==0.5.0
sentence-transformers==2.7.0
langdetect==1.0.9
httpx==0.27.0
python-dotenv==1.0.1
pytest==8.2.0
ruff==0.4.4
numpy<2.0.0
